Add files via upload
mutation and query integrate
