+++ package.json
@@ -14,6 +14,9 @@
   "dependencies": {
     "apollo-server": "^3.5.0",
     "apollo-server-core": "^3.5.0",
-    "graphql": "^15.7.2"
+    "bcryptjs": "^2.4.3",
+    "graphql": "^15.7.2",
+    "jsonwebtoken": "^9.0.0",
+    "mongoose": "^7.0.0"
   }
 }
